refactor(domain) : Car, Cars 클래스 domain 패키지로 리펙토링
