+++ src/test/java/racingCar/domain/CarsTest.java
@@ -1,8 +1,10 @@
-package racingCar;
+package racingCar.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingCar.domain.Car;
+import racingCar.domain.Cars;
 
 import java.util.ArrayList;
 import java.util.List;
